GPII-147: don't convert model on init
No longer converting the model on init. What was happening is that due to the model transformation layer, if a change was saved and the page reloaded, it would be set and unset. This is because the model value for the enactor would be set correctly on init, but the preset models would then be set and transform the enactor models one at a time. To get around this, the modelChange listener isn't bound until after all of the enactors are ready.