GPII-163: Adjusted the document order of the DT
Moved the Discovery Tool higher up in the DOM order to implement the desired tab order. This couldn't be achieved with straight tabindex values since UIO uses an iframe and the tab indexes of this would be calculated separate from the parent document. This would have caused the tab focus to jump back to the top of the page after tabbing away from the open/close button on the panel.

Also removed a call to fluid("tabable") as it was unneeded and caused issues in firefox when trying to set tabindexes of 1 or higher.