GPII-549: Save on transition to full editor
Implemented a workflow by which on clicking the "Full Editor" link, the preferences are saved and the user is sent back to the PMT. To accomplish this the gpiiStore's set method had to be refactored to take in an event. Transitioning to the PMT will only occur if a pageTransitionRequest (clicking on the "Full Editor Link") and a save operation have occurred.